Created Unit Tests
Let's imagine, we have a BarberShop class that has two methods: addCustomer and removeCustomer. These methods add and remove Customer objects from the shop's list of customers.

The testAddCustomer method creates a new BarberShop instance and a new Customer object, adds the customer to the shop, and then checks that the customer count is equal to 1.

The testRemoveCustomer method follows a similar process, but after adding the customer to the shop, it then removes the customer and checks that the customer count is equal to 0.

These tests ensure that the addCustomer and removeCustomer methods are working as expected and correctly modifying the shop's list of customers.