Created Integration test
Let's imagine we have a BarberShop class that has three methods: addCustomer, getCustomerCount, and getNextCustomer. These methods add a Customer object to the shop, return the number of customers in the shop, and return the next customer in the shop's queue, respectively.

The testAddCustomerAndGetCustomerCount method creates a new BarberShop instance and a new Customer object, adds the customer to the shop, gets the customer count, and then checks that the customer count is equal to 1.

The testAddCustomerAndGetNextCustomer method follows a similar process, but adds two customers to the shop, and then gets the next customer in the queue. The test checks that the next customer is equal to the first customer added.

These tests ensure that the integration between the addCustomer, getCustomerCount, and getNextCustomer methods is working as expected and correctly adding and returning customers in the shop.